api/requirement.txt
tensorflow==2.16.1
fastapi
streamlit
uvicorn
opencv-python
python-multipart
pillow
tensorflow-serving-api==2.16.1
matplotlib
numpy